# app/lib/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "FavoriteBatchCheckRequest": {
        "type": "object",
        "description": "Batch favorite check request.",
        "required": ["article_ids"],
        "properties": {
          "article_ids": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ArticleId"
            },
            "description": "Article identifiers to check.",
            "maxItems": 500
          },
          "db_name": {
            "type": "string",
            "description": "Source index database name.",
            "maxLength": 255
          }
        }
      },
...
      "FavoriteCheckResponse": {
        "type": "object",
        "description": "Favorite membership row returned by check endpoints.",
        "required": ["folder_id", "folder_name"],
        "properties": {
          "folder_id": {
            "type": "integer",
            "format": "int64",
            "description": "Folder row identifier."
          },
          "folder_name": {
            "type": "string",
            "description": "Folder display name."
          }
        }
      },
...
      "FolderCreate": {
        "type": "object",
        "description": "Create-folder request payload.",
        "required": ["name"],
        "properties": {
          "is_tracking": {
            "type": "boolean",
            "description": "Whether the folder should become the user's tracking folder."
          },
          "name": {
            "type": "string",
            "description": "Folder display name.",
            "maxLength": 100,
            "minLength": 1
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer"
      },
      "session_cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "litradar_session"
      }
    }
  },
  "info": {
    "title": "LitRadar API",
    "description": "Generated OpenAPI document for the Rust API.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/favorites/check": {
      "get": {
        "tags": ["favorites"],
        "summary": "Check which folders contain an article.",
        "operationId": "check_favorite",
        "parameters": [
          {
            "name": "article_id",
            "in": "query",
            "description": "Article identifier.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "db_name",
            "in": "query",
            "description": "Source database name.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Favorite folder memberships.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FavoriteCheckResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/favorites/check/batch": {
      "post": {
        "tags": ["favorites"],
        "summary": "Batch check which folders contain articles.",
        "operationId": "check_favorites_batch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FavoriteBatchCheckRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Batch favorite folder memberships.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FavoriteBatchCheckResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/favorites/folders": {
      "get": {
        "tags": ["favorites"],
        "summary": "List all folders for the authenticated user.",
        "operationId": "list_folders",
        "responses": {
          "200": {
            "description": "Favorite folders.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FolderResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "session_cookie": []
          }
        ]
      },
      "post": {
        "tags": ["favorites"],
        "summary": "Create a new folder.",
        "operationId": "create_folder",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FolderCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Created favorite folder.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FolderResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/favorites/folders/{folder_id}": {
      "delete": {
        "tags": ["favorites"],
        "summary": "Delete an existing folder.",
        "operationId": "delete_folder",
        "parameters": [
          {
            "name": "folder_id",
            "in": "path",
            "description": "Folder row identifier.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Folder deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "400": {
            "description": "The tracking folder is required by active PushPlus synchronization.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "session_cookie": []
          }
        ]
      },
      "put": {
        "tags": ["favorites"],
        "summary": "Rename an existing folder.",
        "operationId": "rename_folder",
        "parameters": [
          {
            "name": "folder_id",
            "in": "path",
            "description": "Folder row identifier.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FolderRename"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Folder renamed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/favorites/folders/{folder_id}/articles": {
      "get": {
        "tags": ["favorites"],
        "summary": "List favorited articles in a folder.",
        "operationId": "list_folder_articles",
        "parameters": [
          {
            "name": "folder_id",
            "in": "path",
            "description": "Folder row identifier.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum row count.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Offset row count.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Favorite articles.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FavoriteArticleResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "session_cookie": []
          }
        ]
      },
      "post": {
        "tags": ["favorites"],
        "summary": "Add one favorite to a folder.",
        "operationId": "add_favorite",
        "parameters": [
          {
            "name": "folder_id",
            "in": "path",
            "description": "Folder row identifier.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FavoriteAdd"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Created favorite row.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FavoriteResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/favorites/folders/{folder_id}/articles/bulk": {
      "post": {
        "tags": ["favorites"],
        "summary": "Bulk add favorites to a folder.",
        "operationId": "bulk_add",
        "parameters": [
          {
            "name": "folder_id",
            "in": "path",
            "description": "Folder row identifier.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FavoriteBulkAdd"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Bulk favorite add result.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FavoriteBulkAddResult"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/favorites/folders/{folder_id}/articles/bulk-move": {
      "post": {
        "tags": ["favorites"],
        "summary": "Bulk move favorites between folders.",
        "operationId": "bulk_move",
        "parameters": [
          {
            "name": "folder_id",
            "in": "path",
            "description": "Folder row identifier.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FavoriteBulkMove"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Bulk favorite move result.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FavoriteBulkResult"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/favorites/folders/{folder_id}/articles/bulk-remove": {
      "post": {
        "tags": ["favorites"],
        "summary": "Bulk remove favorites from a folder.",
        "operationId": "bulk_remove",
        "parameters": [
          {
            "name": "folder_id",
            "in": "path",
            "description": "Folder row identifier.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FavoriteBulkRemove"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Bulk favorite remove result.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FavoriteBulkResult"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/favorites/folders/{folder_id}/articles/page": {
      "get": {
        "tags": ["favorites"],
        "summary": "Return one stable cursor page from an owned favorite folder.",
        "operationId": "list_folder_article_page",
        "parameters": [
          {
            "name": "folder_id",
            "in": "path",
            "description": "Folder row identifier.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum row count, defaulting to fifty.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "maximum": 500,
              "minimum": 1
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Opaque continuation bound to this user and folder.",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 128
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Stable favorite article page.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FavoriteArticlePage"
                }
              }
            }
          },
          "400": {
            "description": "Invalid limit or cursor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Owned folder not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/favorites/folders/{folder_id}/articles/{article_id}": {
      "delete": {
        "tags": ["favorites"],
        "summary": "Remove one favorite from a folder.",
        "operationId": "remove_favorite",
        "parameters": [
          {
            "name": "folder_id",
            "in": "path",
            "description": "Folder row identifier.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "article_id",
            "in": "path",
            "description": "Article identifier.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "db_name",
            "in": "query",
            "description": "Source database name.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Favorite removed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/favorites/folders/{folder_id}/count": {
      "get": {
        "tags": ["favorites"],
        "summary": "Get the favorite count for a folder.",
        "operationId": "folder_count",
        "parameters": [
          {
            "name": "folder_id",
            "in": "path",
            "description": "Folder row identifier.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Favorite count.",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/favorites/folders/{folder_id}/export": {
      "get": {
        "tags": ["favorites"],
        "summary": "Export one folder's favorites in a citation format.",
        "operationId": "export_folder",
        "parameters": [
          {
            "name": "folder_id",
            "in": "path",
            "description": "Folder row identifier.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "format",
            "in": "query",
            "description": "Export format.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Complete citation export download."
          },
          "400": {
            "description": "The export format is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "The favorite folder does not exist for the user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "413": {
            "description": "The item or final UTF-8 byte limit was exceeded; no partial export is returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Citation metadata could not be read safely.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/favorites/tracking": {
      "get": {
        "tags": ["favorites"],
        "summary": "Get the current tracking folder for the user.",
        "operationId": "get_tracking",
        "responses": {
          "200": {
            "description": "Current favorite tracking folder.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FavoriteTrackingResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "session_cookie": []
          }
        ]
      },
      "put": {
        "tags": ["favorites"],
        "summary": "Set a folder as the current tracking folder.",
        "operationId": "set_tracking",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrackingSetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Tracking folder updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "session_cookie": []
          }
        ]
      }
    },
...
  },
...
}